AHP Fisnished

+++ MultiCriteraDecision/Helper/Consistency.cs
@@ -10,14 +10,14 @@ namespace MultiCriteriaDecision.Helper
         //saaty 
         private const double ReferenceConsistencyRatio= 0.10;
 
-        public static double Check(Matrix ComparisonMatrix, Matrix WeigtMatrix)
+        public static float Check(Matrix ComparisonMatrix, Matrix WeigtMatrix)
         {
             int tmp_Count = WeigtMatrix.RowCount;
-            double tmp_MaxEigen = ComparisonMatrix.GetMaxEigenvalueByWeight(WeigtMatrix);
-            double tmp_ConsistencyRatio = (tmp_MaxEigen - tmp_Count) / (tmp_Count - 1);
-            double tmp_RandomIndex = DecisionDictionary.ConsistencyRatio[tmp_Count];
+            float tmp_MaxEigen = (float)ComparisonMatrix.GetMaxEigenvalueByWeight(WeigtMatrix);
+            float tmp_ConsistencyRatio = (tmp_MaxEigen - tmp_Count) / (tmp_Count - 1);
+            float tmp_RandomIndex = DecisionDictionary.ConsistencyRatio[tmp_Count];
             tmp_ConsistencyRatio = tmp_ConsistencyRatio / tmp_RandomIndex;
-            return tmp_ConsistencyRatio;
+            return (float)tmp_ConsistencyRatio;
         }
     }
 }
